refactor(samples): incorporated review comments
