--- samples/snippets/site_key.py
@@ -1,213 +1,0 @@
-# Copyright 2021 Google Inc. All Rights Reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#      http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-# [START recaptcha_enterprise_create_site_key]
-# [START recaptcha_enterprise_get_site_key]
-# [START recaptcha_enterprise_list_site_keys]
-# [START recaptcha_enterprise_update_site_key]
-# [START recaptcha_enterprise_delete_site_key]
-import time
-
-from google.cloud import recaptchaenterprise_v1
-
-
-# [END recaptcha_enterprise_delete_site_key]
-# [END recaptcha_enterprise_update_site_key]
-# [END recaptcha_enterprise_list_site_keys]
-# [END recaptcha_enterprise_get_site_key]
-# [END recaptcha_enterprise_create_site_key]
-
-
-# [START recaptcha_enterprise_create_site_key]
-def create_site_key(project_id: str, domain_name: str) -> str:
-    """Create reCAPTCHA Site key which binds a domain name to a unique key.
-    Args:
-    project_id : GCloud Project ID.
-    domain_name: Specify the domain name in which the reCAPTCHA should be activated.
-    """
-    client = recaptchaenterprise_v1.RecaptchaEnterpriseServiceClient()
-
-    # Set the type of the reCAPTCHA to be displayed.
-    # For different types, see: https://cloud.google.com/recaptcha-enterprise/docs/keys
-    web_settings = recaptchaenterprise_v1.WebKeySettings()
-    web_settings.allowed_domains.append(domain_name)
-    web_settings.allow_amp_traffic = False
-    web_settings.integration_type = web_settings.IntegrationType.SCORE
-
-    key = recaptchaenterprise_v1.Key()
-    key.display_name = "any descriptive name for the key"
-    key.web_settings = web_settings
-
-    # Create the request.
-    request = recaptchaenterprise_v1.CreateKeyRequest()
-    request.parent = f'projects/{project_id}'
-    request.key = key
-
-    # Get the name of the created reCAPTCHA site key.
-    response = client.create_key(request)
-    recaptcha_site_key = response.name.rsplit("/", maxsplit=1)[1]
-    print("reCAPTCHA Site key created successfully. Site Key: " + recaptcha_site_key)
-    return recaptcha_site_key
-
-
-# [END recaptcha_enterprise_create_site_key]
-
-
-# [START recaptcha_enterprise_get_site_key]
-def get_key(project_id: str, recaptcha_site_key: str) -> None:
-    """
-    Get the reCAPTCHA site key present under the project ID.
-
-    Args:
-    project_id: GCloud Project ID.
-    recaptcha_site_key: Specify the site key to get the details.
-    """
-
-    client = recaptchaenterprise_v1.RecaptchaEnterpriseServiceClient()
-
-    # Construct the key details.
-    key_name = f'projects/{project_id}/keys/{recaptcha_site_key}'
-
-    request = recaptchaenterprise_v1.GetKeyRequest()
-    request.name = key_name
-
-    key = client.get_key(request)
-    print("Successfully obtained the key !" + key.name)
-
-
-# [END recaptcha_enterprise_get_site_key]
-
-
-# [START recaptcha_enterprise_list_site_keys]
-def list_site_keys(project_id: str) -> None:
-    """ List all keys present under the given project ID.
-
-    Args:
-    project_id: GCloud Project ID.
-    """
-
-    project_name = f'projects/{project_id}'
-
-    client = recaptchaenterprise_v1.RecaptchaEnterpriseServiceClient()
-
-    # Set the project id to list the keys present in it.
-    request = recaptchaenterprise_v1.ListKeysRequest()
-    request.parent = project_name
-
-    response = client.list_keys(request)
-    print("Listing reCAPTCHA site keys: ")
-    for i, key in enumerate(response):
-        print(str(i) + '. ' + key.name)
-
-
-# [END recaptcha_enterprise_list_site_keys]
-
-
-# [START recaptcha_enterprise_update_site_key]
-def update_site_key(project_id: str, recaptcha_site_key: str, domain_name: str) -> None:
-    """
-    Update the properties of the given site key present under the project id.
-
-    Args:
-    project_id: GCloud Project ID.
-    recaptcha_site_key: Specify the site key.
-    domain_name: Specify the domain name for which the settings should be updated.
-    """
-
-    client = recaptchaenterprise_v1.RecaptchaEnterpriseServiceClient()
-
-    # Construct the key details.
-    key_name = f'projects/{project_id}/keys/{recaptcha_site_key}'
-
-    # Set the name and the new settings for the key.
-    web_settings = recaptchaenterprise_v1.WebKeySettings()
-    web_settings.allow_amp_traffic = True
-    web_settings.allowed_domains.append(domain_name)  # TODO
-
-    key = recaptchaenterprise_v1.Key()
-    key.name = key_name
-    key.web_settings = web_settings
-
-    update_key_request = recaptchaenterprise_v1.UpdateKeyRequest()
-    update_key_request.key = key
-    client.update_key(update_key_request)
-
-    time.sleep(5)
-
-    # Retrieve the key and check if the property is updated.
-    get_key_request = recaptchaenterprise_v1.GetKeyRequest()
-    get_key_request.name = key_name
-    response = client.get_key(get_key_request)
-    web_settings = response.web_settings
-
-    # Get the changed property.
-    if not web_settings.allow_amp_traffic:
-        print("Error! reCAPTCHA Site key property hasn't been updated. Please try again !")
-    else:
-        print("reCAPTCHA Site key successfully updated ! ")
-
-
-# [END recaptcha_enterprise_update_site_key]
-
-
-# [START recaptcha_enterprise_delete_site_key]
-def delete_site_key(project_id: str, recaptcha_site_key: str) -> None:
-    """ Delete the given reCAPTCHA site key present under the project ID.
-
-    Args:
-    project_id : GCloud Project ID.
-    recaptcha_site_key: Specify the key ID to be deleted.
-    """
-
-    client = recaptchaenterprise_v1.RecaptchaEnterpriseServiceClient()
-
-    # Construct the key details.
-    key_name = f'projects/{project_id}/keys/{recaptcha_site_key}'
-
-    # Set the project ID and reCAPTCHA site key.
-    request = recaptchaenterprise_v1.DeleteKeyRequest()
-    request.name = key_name
-
-    client.delete_key(request)
-    print("reCAPTCHA Site key deleted successfully ! ")
-
-
-# [END recaptcha_enterprise_delete_site_key]
-
-
-def main(default_project_id: str, domain_name: str) -> str:
-    recaptcha_site_key = create_site_key(project_id=default_project_id, domain_name=domain_name)
-    list_site_keys(project_id=default_project_id)
-    get_key(project_id=default_project_id, recaptcha_site_key=recaptcha_site_key)
-    update_site_key(project_id=default_project_id, recaptcha_site_key=recaptcha_site_key, domain_name=domain_name)
-    delete_site_key(project_id=default_project_id, recaptcha_site_key=recaptcha_site_key)
-
-    return recaptcha_site_key
-
-
-if __name__ == "__main__":
-    import google.auth
-    import google.auth.exceptions
-
-    # TODO(developer): Replace the below variables before running
-    try:
-        default_project_id = google.auth.default()[1]
-        domain_name = "localhost"
-    except google.auth.exceptions.DefaultCredentialsError:
-        print(
-            "Please use `gcloud auth application-default login` "
-            "or set GOOGLE_APPLICATION_CREDENTIALS to use this script."
-        )
-    else:
-        main(default_project_id, domain_name)
